+++ translator/src/test/scala/soda/lib/OptionSDSpec.scala
@@ -7,7 +7,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   import scala.util.Failure
 
   test ("should test an empty option") {
-    lazy val empty = NoInst ()
+    lazy val empty = NoElem ()
 
     assert (empty.isEmpty &&
       ! empty.isDefined &&
@@ -15,7 +15,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("should test a non empty option") {
-    lazy val element = SomeInst (1 )
+    lazy val element = SomeElem (1 )
 
     assert (! element.isEmpty &&
       element.isDefined &&
@@ -23,7 +23,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("should get a default value, when empty") {
-    lazy val empty = NoInst [Int]  ()
+    lazy val empty = NoElem [Int]  ()
     lazy val expected = 1
     lazy val obtained = empty.getOrElse (1 )
 
@@ -31,7 +31,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("should get a value") {
-    lazy val empty = SomeInst [Int]  (2 )
+    lazy val empty = SomeElem [Int]  (2 )
     lazy val expected = 2
     lazy val obtained = empty.getOrElse (1 )
 
@@ -41,7 +41,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   test ("should open an empty option") {
     lazy val result_if_empty: String = "It is empty."
     def result_if_non_empty (value: String ): String = "Its value is " + value + "."
-    lazy val empty = NoInst [String]  ()
+    lazy val empty = NoElem [String]  ()
     lazy val expected = "It is empty."
     lazy val obtained = empty.opt (ifEmpty = result_if_empty, ifNonEmpty = result_if_non_empty )
 
@@ -51,7 +51,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   test ("should open an non empty option") {
     lazy val result_if_empty: String = "It is empty."
     def result_if_non_empty (value: String ): String = "Its value is " + value + "."
-    lazy val some_element = SomeInst [String]  ("0")
+    lazy val some_element = SomeElem [String]  ("0")
     lazy val expected = "Its value is 0."
     lazy val obtained = some_element.opt (ifEmpty = result_if_empty, ifNonEmpty = result_if_non_empty )
 
@@ -61,7 +61,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   test ("should try fold an empty option") {
     lazy val result_if_empty: String = "It is empty."
     def result_if_non_empty (value: String ): String = "Its value is " + value + "."
-    lazy val empty = NoInst [String]  ()
+    lazy val empty = NoElem [String]  ()
     lazy val expected = "It is empty."
     lazy val obtained = empty.fold (ifEmpty = result_if_empty, f = result_if_non_empty )
 
@@ -71,7 +71,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   test ("should try fold an non empty option") {
     lazy val result_if_empty: String = "It is empty."
     def result_if_non_empty (value: String ): String = "Its value is " + value + "."
-    lazy val some_element = SomeInst [String]  ("0")
+    lazy val some_element = SomeElem [String]  ("0")
     lazy val expected = "Its value is 0."
     lazy val obtained = some_element.fold (ifEmpty = result_if_empty, f = result_if_non_empty )
 
@@ -80,8 +80,8 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should map empty to empty") {
     def to_string (n: Int ): String = "" + n
-    lazy val empty = NoInst [Int]  ()
-    lazy val expected = NoInst [String]  ()
+    lazy val empty = NoElem [Int]  ()
+    lazy val expected = NoElem [String]  ()
     lazy val obtained = empty.map (to_string )
 
     assert (obtained == expected )
@@ -89,26 +89,26 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
 
   test ("should map a non-empty to another non-empty") {
     def to_string (n: Int ): String = "" + n
-    lazy val some_element = SomeInst [Int]  (2 )
-    lazy val expected = SomeInst [String]  ("2")
+    lazy val some_element = SomeElem [Int]  (2 )
+    lazy val expected = SomeElem [String]  ("2")
     lazy val obtained = some_element.map (to_string )
 
     assert (obtained == expected )
   }
 
   test ("should flat map empty to empty") {
-    def to_string (n: Int ): SomeInst [String] = SomeInst [String]  ("" + n )
-    lazy val empty = NoInst [Int]  ()
-    lazy val expected = NoInst [String]  ()
+    def to_string (n: Int ): SomeElem [String] = SomeElem [String]  ("" + n )
+    lazy val empty = NoElem [Int]  ()
+    lazy val expected = NoElem [String]  ()
     lazy val obtained = empty.flatMap (to_string )
 
     assert (obtained == expected )
   }
 
   test ("should flat map a non-empty to another non-empty") {
-    def to_string (n: Int ): SomeInst [String] = SomeInst [String]  ("" + n )
-    lazy val some_element = SomeInst [Int]  (2 )
-    lazy val expected = SomeInst [String]  ("2")
+    def to_string (n: Int ): SomeElem [String] = SomeElem [String]  ("" + n )
+    lazy val some_element = SomeElem [Int]  (2 )
+    lazy val expected = SomeElem [String]  ("2")
     lazy val obtained = some_element.flatMap (to_string )
 
     assert (obtained == expected )
@@ -128,18 +128,18 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
     lazy val maybeB = toInt (stringB )
     lazy val maybeC = toInt (stringC )
 
-    lazy val expected = SomeInst (6 )
+    lazy val expected = SomeElem (6 )
     lazy val obtained =
-      maybeA.opt (ifEmpty = NoInst, ifNonEmpty = a =>
-          maybeB.opt (ifEmpty = NoInst, ifNonEmpty = b =>
-              maybeC.opt (ifEmpty = NoInst, ifNonEmpty = c =>
-                  SomeInst (a + b + c ) ) ) )
+      maybeA.opt (ifEmpty = NoElem, ifNonEmpty = a =>
+          maybeB.opt (ifEmpty = NoElem, ifNonEmpty = b =>
+              maybeC.opt (ifEmpty = NoElem, ifNonEmpty = c =>
+                  SomeElem (a + b + c ) ) ) )
 
     assert (obtained == expected )
   }
 
   test ("toOption with non empty option") {
-    lazy val input: OptionSD [Int] = SomeInst (1 )
+    lazy val input: OptionSD [Int] = SomeElem (1 )
     lazy val expected: Option [Int] = Some (1 )
     lazy val obtained = input.toOption
 
@@ -147,7 +147,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("toOption with another non empty option") {
-    lazy val input: SomeInst [Int] = SomeInst (2 )
+    lazy val input: SomeElem [Int] = SomeElem (2 )
     lazy val expected: Some [Int] = Some (2 )
     lazy val obtained = input.toOption
 
@@ -155,7 +155,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("toOption with empty option") {
-    lazy val input: OptionSD [Int] = NoInst ()
+    lazy val input: OptionSD [Int] = NoElem ()
     lazy val expected = None
     lazy val obtained = input.toOption
 
@@ -163,7 +163,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("toSeq with non empty option") {
-    lazy val input: OptionSD [Int] = SomeInst (1 )
+    lazy val input: OptionSD [Int] = SomeElem (1 )
     lazy val expected: Seq [Int] = Seq (1 )
     lazy val obtained = input.toSeq
 
@@ -171,7 +171,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("toSeq with another non empty option") {
-    lazy val input: SomeInst [Int] = SomeInst (2 )
+    lazy val input: SomeElem [Int] = SomeElem (2 )
     lazy val expected: Seq [Int] = Seq (2 )
     lazy val obtained = input.toSeq
 
@@ -179,7 +179,7 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("toSeq with empty option") {
-    lazy val input: OptionSD [Int] = NoInst ()
+    lazy val input: OptionSD [Int] = NoElem ()
     lazy val expected = Seq ()
     lazy val obtained = input.toSeq
 
@@ -187,24 +187,24 @@ case class OptionSDSpec () extends org.scalatest.funsuite.AnyFunSuite {
   }
 
   test ("filter should work for None") {
-    lazy val input = NoInst [Int]  ()
-    lazy val expected = NoInst [Int]  ()
+    lazy val input = NoElem [Int]  ()
+    lazy val expected = NoElem [Int]  ()
     lazy val obtained = input.filter (x => true )
 
     assert (obtained == expected )
   }
 
   test ("filter should work for Some, if predicate does not hold") {
-    lazy val input = SomeInst [Int]  (0 )
-    lazy val expected = NoInst [Int]  ()
+    lazy val input = SomeElem [Int]  (0 )
+    lazy val expected = NoElem [Int]  ()
     lazy val obtained = input.filter (x => x > 0 )
 
     assert (obtained == expected )
   }
 
   test ("filter should work for Some, if predicate holds") {
-    lazy val input = SomeInst [Int]  (1 )
-    lazy val expected = SomeInst [Int]  (1 )
+    lazy val input = SomeElem [Int]  (1 )
+    lazy val expected = SomeElem [Int]  (1 )
     lazy val obtained = input.filter (x => x > 0 )
 
     assert (obtained == expected )